requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
jinja2==3.1.3
aiofiles==23.2.1
gunicorn==21.2.0 